+++ NEWS.md
@@ -1,3 +1,6 @@
+# Version 0.4.0
+* Added PARC-connectedness calculation to the `connectedness()` function.
+
 # Version 0.3.0
 * Fixed the code to accept raster integer data type.
 * Added a scaling factor (`scale` arg) to convert condition data to the 0–1 range.